2007-01-02  Michael McCarty <mkmccarty@gmail.com>
	* V39

	* fixed some unitialized values

	* replaced calls to kill() with remove_savedat()

2007-01-02  Pace Willisson  <pace@rhino.pacew.com>

	* V38

	* rpm file generation

	* V37

	* allow videos to be skipped when viewing missions

	* V36

	* make little screens work during launch

	* V35

	* remove stubs that are no longer needed (now unimp() is gone);
	remove modem and play by mail stuff until everything else works

	* V34

	* hack around messed up crane image (search for fixing sImg) in
	port.c

	* V33

	* fix some NULL pointer problems in mc.c and prest.c

	* V32

	* enable gcc warning related to structured initializers

	* V31

	* enable gcc warning about comparing signed and unsigned values

	* V30

	* enable gcc check for use of uninitialized local
	variable - put in lots of comments like this:

	 spc[0] = 0; /* XXX check uninitialized */

	 Some of these are real bugs, where there was no way to initialize
	the variable before it was used.  In other cases, the code path
	that gcc found to an uninitialized use may be impossible to
	traverse given other data constraints.  Future project:
	look for all these comments and either fix any bugs present,
	or remove the comment if the code is ok.

	* V29

	* add extra braces and parens so gcc's Wparentheses warning
	will be useful.  I compiled the program before and after
	these changes with no difference in the .o files, so the
	explicit groups is now what the compiler has done all along.
	there were two paces in future.c, marked with XXX, where
	the indenting suggested that the other order was intended -
	this is worth checking

	* V28

	* the ai was looping forever in RDafford because roll was
	negative due to Manned Safety being more than MaxRD.  the
	dos game survived this because roll was a 16 bit number,
	so it didn't take to long to count it down until it wrapped
	to 0.  but it was never going to finish with a 32 bit roll value

	for now, delcare roll as i16 to make it behave a least as well as dos...

	remove calls to inp() that checked for vblank

	change WORD to i16 to make Spot data structures work

	enable SPOT_ON visuals.  sound will come later; I've seen some
	helicoptors, but no vehicle assembly yet

	* V27

	* remaining unsued locals

	* V26

	* another block of unused locals

	* V25

	* remove more unused locals

	* V24

	* I messed up trying to apply tag V23 to Michael's changes.
	This check just puts V24 on the current files.

2007-01-02  Michael McCarty <mkmccarty@gmail.com>

	* V23

	* added BARIS_ROOT envorinment variable so GAMEDAT and ROM assets
	can exist on a local disk

	* Removed definition of _ExceptInit(void) which was originally used to
	trim unused Borland C exception code and reduce code size

2007-01-02  Pace Willisson  <pace@rhino.pacew.com>

	* V22

	* remove remaining ems stuff; remove libopus references; fix
	credits; re-enable water flowing

	* V21

	* enable mission failures (and make command line arg to force all
	to succeed again); allow jumbled arg order in draw/clear rectangle;
	give symbolic names to xMODE flags (for cloud testing); failure
	animation; data type size problem reading NTABLE.DAT;

	* V20

	* remove deadwood

	* V19

	* a few notes in README

	* V18

	* improve finding files on cdrom; move savedat to ~/.raceintospace;
	put music files in /usr/share/raceintospace/music

2007-01-01  Pace Willisson  <pace@rhino.pacew.com>

	* V17

	* clean up timers and frame rate; improve mouse click handling;
	remove play by mail and modem choices; make flag waving and
	water flowing work;

	* V16

	* remove unused files

	* V15

	* add music

	* V14

	* add audio to mission playback

	* V13

	* make mission display work; fix up inter-frame delays in video;

	* V12

	* switch from gtk to sdl; make use of toupper safe against
	values outside of unsigned char range; put some sleeps
	in when waiting for input

	* V11

	* fix display in VAB; some other helper programs

	* V10

	* more pathname details

	* V9

	* pathname handling

	* V8

	* PlaySequence, Replay

	* change "kill()" to "remove_savegame()"

	* remove uses of SetROMDir, chdir, unlink

2006-12-31  Pace Willisson  <pace@rhino.pacew.com>

	* V7

	* old bug: initialization of FList when looking for saved games

	* frm decoding; first page of mission history screen (with
	still image from movie)

2006-12-30  Pace Willisson  <pace@rhino.pacew.com>

	* V6

	* change many RLED to RLED_img

	* avoid segfaults due to NULL pointers

2006-12-29  Pace Willisson  <pace@rhino.pacew.com>

	* V5

	* V4

2006-12-28  Pace Willisson  <pace@rhino.pacew.com>

	* V3

	* V2

	* V1

+++ main.c
@@ -422,7 +422,7 @@ int main(int argc, char *argv[])
     else fin=sOpen("RAST.DAT","rb",0);
     i=(int) filelength(fileno(fin));  // get length
     fread((char *)buffer,i,1,fin); fclose(fin);
-    printf ("reading Players: size = %d\n", sizeof (struct Players));
+    printf ("reading Players: size = %d\n", (int)sizeof (struct Players));
     RLED(buffer,(char *)Data,i);
     if (Data->Checksum!=(sizeof (struct Players))) {
       printf("BARIS Note: Wrong version of Data File.\n");
